+++ test/models/dataset_test.rb
@@ -9,4 +9,9 @@ class DatasetTest < ActiveSupport::TestCase
   test "should find dataset by slug" do
     assert_equal datasets(:public), Dataset.find_by_param(datasets(:public).slug)
   end
+
+  test "should remove file indexes" do
+    datasets(:public).reset_folder_indexes # Remove and return an array with potential .sleepdata.index files
+    assert_equal [], datasets(:public).reset_folder_indexes # Should find no .sleepdata.index files
+  end
 end